6. HW6 Classes and Objects - Task1
Create a class "PhotoAlbum" with a private attribute "numberOfPages".
It should also have a public method "GetNumberOfPages", which will return the number of pages.
The default constructor will create an album with 16 pages. There will be an additional constructor, with which we can specify the number of pages we want in the album.
Create a class "BigPhotoAlbum" whose constructor will create an album with 64 pages.
In main method create a "PhotoAlbum" instance with its default constructor and one with 24 pages. Also create "BigPhotoAlbum" and show the number of pages that the three albums have.